Improve code quality: split train_base.py, fix lint/type errors, add tests
- Split train_base.py (1,415 LOC) into focused modules:
  - diagnostics.py: DiagnosticsCallback (~160 LOC)
  - evaluation.py: eval_policy, evaluate, record_stage_video (~270 LOC)
  - cli.py: main, _apply_overrides, _cast_value (~200 LOC)
  - train_base.py re-exports all symbols for backward compatibility

- Extract shared reward helpers into BaseDinoEnv:
  - _reward_alive(), _reward_energy(), _reward_action_smoothness()
  - Common sensor layout constants (_sensor_gyro/accel/quat_start)
  - Reduces ~50 lines of duplication across 3 species envs

- Fix ruff lint errors: unused imports in test_sweep.py, unsorted imports
  and f-strings without placeholders in training notebook

- Fix mypy type errors: untyped parameter in reporting.py save_results_json,
  np.savez arg-type in train_base.py

- Add 48 unit tests for reporting.py (was at 9% coverage):
  - format_duration, format_duration_hms
  - write_stage_summary, write_training_summary
  - save_results_json
