+++ environments/trex/envs/trex_env.py
@@ -167,9 +167,8 @@ class TRexEnv(BaseDinoEnv):
         # Sensor indices (order matches MJCF sensor definition)
         # pelvis_gyro(3), pelvis_accel(3), pelvis_orientation(4),
         # r_foot_touch(1), l_foot_touch(1)
-        self._sensor_gyro_start = 0
-        self._sensor_accel_start = 3
-        self._sensor_quat_start = 6
+        # _sensor_gyro_start, _sensor_accel_start, _sensor_quat_start
+        # are inherited from BaseDinoEnv (0, 3, 6 respectively).
         self._sensor_r_foot = 10
         self._sensor_l_foot = 11
 
@@ -241,16 +240,12 @@ class TRexEnv(BaseDinoEnv):
         reward_forward = self.forward_vel_weight * forward_vel_norm
         info["reward_forward"] = reward_forward
 
-        # 2. Alive bonus
-        reward_alive = self.alive_bonus
+        # 2. Alive bonus (shared helper)
+        reward_alive = self._reward_alive()
         info["reward_alive"] = reward_alive
 
-        # 3. Energy penalty (normalized by number of actuators)
-        energy = np.sum(np.square(action))
-        assert self.action_space.shape is not None
-        n_actuators = self.action_space.shape[0]
-        energy_norm = energy / n_actuators
-        reward_energy = -self.energy_penalty_weight * energy_norm
+        # 3. Energy penalty (shared helper)
+        reward_energy = self._reward_energy(action)
         info["reward_energy"] = reward_energy
 
         # 4. Tail stability (penalize high angular velocity at tail tip) — normalized
@@ -353,17 +348,8 @@ class TRexEnv(BaseDinoEnv):
         info["contact_asymmetry"] = contact_asymmetry
         info["reward_gait"] = reward_gait
 
-        # 10. Action smoothness (penalize large action changes between steps)
-        if self._prev_action is not None:
-            action_delta = float(np.sum(np.square(action - self._prev_action)))
-            assert self.action_space.shape is not None
-            max_action_delta = self.action_space.shape[0] * 4.0
-            action_delta_norm = action_delta / max_action_delta
-            reward_smoothness = -self.smoothness_weight * action_delta_norm
-        else:
-            action_delta = 0.0
-            reward_smoothness = 0.0
-        self._prev_action = action.copy()
+        # 10. Action smoothness (shared helper)
+        reward_smoothness, action_delta = self._reward_action_smoothness(action)
         info["action_delta"] = action_delta
         info["reward_smoothness"] = reward_smoothness
